requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blessings==1.7
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==4.0.2
cycler==0.10.0
decorator==4.3.0
defusedxml==0.6.0
docker-pycreds==0.4.0
entrypoints==0.3
gast==0.3.2
gitdb==4.0.2
GitPython==3.1.0
google-pasta==0.1.8
gpustat==0.6.0
GPUtil==1.4.0
gql==0.2.0
graphql-core==1.1
grpcio==1.25.0
h5py==2.10.0
idna==2.9
imageio==2.6.1
importlib-metadata==1.2.0
ipykernel==5.1.3
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.3
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==3.1.2
missingno==0.4.2
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.5
nvidia-ml-py3==7.352.0
olefile==0.46
omegaconf==1.4.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.3.1
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.7
protobuf==3.7.1
psutil==5.6.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytorch-revgrad==0.1.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
sentry-sdk==0.14.2
shap==0.32.1
shortuuid==1.0.1
six==1.12.0
smmap==3.0.1
subprocess32==3.5.4
tensorboard==1.14.0
tensorboardX==1.6
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.40.1
traitlets==4.3.2
urllib3==1.25.8
wandb==0.8.29
watchdog==0.10.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.2
zipp==0.6.0
